# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/amadeus/transfers/search": {
      "post": {
        "operationId": "AmadeusTransfersController_search",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search transfer offers",
        "tags": [
          "Amadeus — Cars & Transfers"
        ]
      }
    },
    "/api/v1/amadeus/transfers/book": {
      "post": {
        "operationId": "AmadeusTransfersController_book",
        "parameters": [
          {
            "name": "offerId",
            "required": true,
            "in": "query",
            "description": "Transfer offer ID from search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Book a transfer",
        "tags": [
          "Amadeus — Cars & Transfers"
        ]
      }
    },
    "/api/v1/amadeus/transfers/cancel": {
      "post": {
        "operationId": "AmadeusTransfersController_cancel",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "query",
            "description": "Transfer order ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "confirmNbr",
            "required": true,
            "in": "query",
            "description": "Transfer confirmation number",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a transfer booking",
        "tags": [
          "Amadeus — Cars & Transfers"
        ]
      }
    },
    "/api/v1/amadeus/insights/most-traveled": {
      "get": {
        "operationId": "AmadeusInsightsController_mostTraveled",
        "parameters": [
          {
            "name": "originCityCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "MAD",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": true,
            "in": "query",
            "description": "YYYY-MM format",
            "schema": {
              "example": "2023-11",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "analytics.travelers.score",
              "type": "string"
            }
          },
          {
            "name": "max",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight most traveled destinations",
        "tags": [
          "Amadeus — Market Insights"
        ]
      }
    },
    "/api/v1/amadeus/insights/most-booked": {
      "get": {
        "operationId": "AmadeusInsightsController_mostBooked",
        "parameters": [
          {
            "name": "originCityCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "NCE",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2023-11",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight most booked destinations",
        "tags": [
          "Amadeus — Market Insights"
        ]
      }
    },
    "/api/v1/amadeus/insights/busiest-period": {
      "get": {
        "operationId": "AmadeusInsightsController_busiestPeriod",
        "parameters": [
          {
            "name": "cityCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "PAR",
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": true,
            "in": "query",
            "description": "YYYY format",
            "schema": {
              "example": "2023",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ARRIVING",
                "DEPARTING"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Flight busiest traveling period",
        "tags": [
          "Amadeus — Market Insights"
        ]
      }
    },
...
    "/api/v1/amadeus/itinerary/trip-purpose": {
      "get": {
        "operationId": "AmadeusItineraryController_tripPurpose",
        "parameters": [
          {
            "name": "originLocationCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "LON",
              "type": "string"
            }
          },
          {
            "name": "destinationLocationCode",
            "required": true,
            "in": "query",
            "schema": {
              "example": "AMS",
              "type": "string"
            }
          },
          {
            "name": "departureDate",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-08-01",
              "type": "string"
            }
          },
          {
            "name": "returnDate",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2025-08-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Predict trip purpose (business vs leisure)",
        "tags": [
          "Amadeus — Itinerary Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "✈️ Flight Booking API",
    "description": "\n      ## Multi-Tenant Flight Booking Platform API\n      \n      A scalable flight booking backend with:\n      - **Multi-tenancy** (B2B & B2C)\n      - **Multi-currency** support (14 currencies)\n      - **Dual payment** processing (Stripe + Paystack)\n      - **Real-time** seat availability\n      - **Admin analytics** and reporting\n      - **Global Travel Integration** (Amadeus Search, Price, Booking)\n      \n      ### Amadeus API Modules\n      This platform integrates the following Amadeus modules:\n      - **Flights**: Live search, pricing, and orders (31 endpoints)\n      - **Hotels**: Search, offers, and bookings (10 endpoints)\n      - **Experiences**: Tours, activities, and city search (4 endpoints)\n      - **Transfers**: Cars and transfers search/booking (3 endpoints)\n      - **Insights**: Market trends and traffic analytics (3 endpoints)\n      - **Itinerary**: Travel predictions and planning (1 endpoint)\n      \n      ### Authentication\n      All protected endpoints require a Bearer JWT token.\n      Use `/auth/login` to obtain tokens.\n      \n      ### Rate Limiting\n      API requests are rate-limited to 100 requests per minute per IP.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}